+++ R/data-cn_iecsc.R
@@ -1,0 +1,14 @@
+#' @title The China Inventory of Existing Chemical Substances and Chemicals
+#' @description Note that this data was extracted from messy PDFs and is not
+#'   complete. In particular, (English) chemical names still need to be
+#'   extracted. For further information on the data sets, please visit the
+#'   official website from which the raw data originated:
+#'
+#'   * https://www.mee.gov.cn/gkml/hbb/bgg/201301/t20130131_245810.htm
+#'
+#' @format A data frame with 37,069 rows and 1 variable:
+#' \describe{
+#'   \item{cas_number}{(Character) Chemical Abstracts Service (CAS) registry
+#'     number}
+#' }
+"cn_iecsc"