+++ R/data-jp_cscl_ecs.R
@@ -1,0 +1,26 @@
+#' @title The Japanese Chemical Substances Control Law Existing Chemical
+#'   Substances Data Set
+#' @description Substances that were manufactured or imported for business
+#'   purpose at the time of the promulgation of the CSCL (excluding substances
+#'   that were manufactured or imported for testing and research purposes and
+#'   those that were manufactured or imported for use as reagents), and whose
+#'   names were published pursuant to the CSCL (chemical substances were listed
+#'   in the list of Existing Chemical Substances).
+#'
+#'   Under the current CSCL, substances are designated as General Chemical
+#'   Substances in accordance with paragraph (7) of Article 2 (excluding
+#'   Priority Assessment Chemical Substances, Monitoring Chemical Substances
+#'   and Class I/II Specified Chemical Substances). If a substance corresponds
+#'   to the "Existing Chemical substances", the following items are displayed.
+#' @format A data frame with 56,975 rows and 5 variables:
+#' \describe{
+#'   \item{no}{(Integer) List number for each substance}
+#'   \item{chrip_id}{(Character) The CHemical Risk Information Platform (CHRIP)
+#'     ID}
+#'   \item{cas_rn}{(Character) Chemical Abstracts Service (CAS) registry number}
+#'   \item{cas_identity}{(Character) Additional information about the substance}
+#'   \item{chemical_substance_name}{(Character) The provided name for the
+#'     substance}
+#' }
+#' @source \url{https://www.nite.go.jp/en/chem/chrip/chrip_search/sltLst}
+"jp_cscl_ecs"