+++ R/data-tw_csi_add.R
@@ -1,0 +1,15 @@
+#' @title The Taiwan Chemical Substance Inventory Data Set Addition
+#' @description Note that this data was extracted from image PDFs via optical
+#'   character recognition (OCR) and is likely not 100% complete. The entire
+#'   procedure is documented in the "cleanventory" R package. For further
+#'   information on the data set, please visit the two official websites from
+#'   which the raw data originated:
+#'
+#'   * https://gazette.nat.gov.tw/egFront/detail.do?metaid=78617&log=detailLog
+#'
+#' @format A data frame with 6,230 rows and 1 variable:
+#' \describe{
+#'   \item{cas_or_serial_no}{(Character) Chemical Abstracts Service (CAS)
+#'     registry number or serial number}
+#' }
+"tw_csi_add"