+++ NEWS.md
@@ -1,3 +1,18 @@
+# cleanventory.data 0.3.1 (2022-07-19)
+
+* All data sets now have a two-letter code indicating the country/region of 
+  origin.
+* If necessary, existing data sets were re-separated to reflect individual 
+  origin files. This was relevant for the EU (`eu_clp_*`) and 
+  Taiwan (`tw_csi*`).
+* Added the following national/regional inventories: Philippines (`ph_iccs`, 
+  `ph_emb_mc_07`, `ph_emb_mc_08`), Japan (`jp_cscl_ecs`, `jp_cscl_nacs_1`, 
+  `jp_cscl_nacs_2`), Canada (`ca_dsl`), and China (`cn_iecsc`).
+
+# cleanventory.data 0.2.4 (2022-XX-XX)
+
+* Included data set for `tcsi`.
+
 # cleanventory.data 0.2.3 (2022-05-04)
 
 * Included data set for `ncis`.